move crypto/ecies to erigon-lib
